[FLINK-16498] connect PostgresCatalog to table planner
