+++ flink-connectors/flink-jdbc/src/main/java/org/apache/flink/api/java/io/jdbc/catalog/PostgresTablePath.java
@@ -64,13 +64,9 @@ public class PostgresTablePath {
 		return String.format("%s.%s", pgSchemaName, pgTableName);
 	}
 
-	public String getFullPathWithQuotes() {
-		return String.format("`%s.%s`", pgSchemaName, pgTableName);
-	}
-
 	@Override
 	public String toString() {
-		return getFullPathWithQuotes();
+		return getFullPath();
 	}
 
 	@Override
